requirements.txt
Pillow==10.0.0
stegano==0.8.1
PyQt5==5.15.9